Refactor Ids to shared location
